+++ src/components/Grid/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./Grid";
+export type { Props as GridProps, ItemProps as GridItemProps } from "./Grid.types";