+++ src/components/Popover/Popover.tsx
@@ -1,5 +1,5 @@
 import { classNames } from "utilities/helpers";
-import Flyout, { useFlyoutContext } from "components/_private/Flyout";
+import Flyout, { useFlyoutContext, type FlyoutProps } from "components/_private/Flyout";
 import Dismissible, { type DismissibleProps } from "components/Dismissible";
 import type * as T from "./Popover.types";
 import s from "./Popover.module.css";
@@ -7,23 +7,11 @@ import getPaddingStyles from "styles/padding";
 
 const Popover = (props: T.Props) => {
 	const {
-		id,
-		forcePosition,
-		onOpen,
-		onClose,
-		active,
-		defaultActive,
-		children,
 		width,
-		contentGap,
 		variant = "elevated",
 		triggerType = "click",
 		position = "bottom",
-		disableHideAnimation,
-		disableContentHover,
-		disableCloseOnOutsideClick,
-		instanceRef,
-		containerRef,
+		...flyoutProps
 	} = props;
 	const padding = props.padding ?? (variant === "headless" ? 0 : 4);
 	const trapFocusMode =
@@ -37,29 +25,15 @@ const Popover = (props: T.Props) => {
 	);
 
 	return (
-		// @ts-ignore
 		<Flyout
-			id={id}
-			instanceRef={instanceRef}
+			{...(flyoutProps as FlyoutProps)}
 			position={position}
-			forcePosition={forcePosition}
-			onOpen={onOpen}
-			onClose={onClose}
 			trapFocusMode={trapFocusMode}
 			triggerType={triggerType}
-			active={active}
-			defaultActive={defaultActive}
 			width={width}
-			disableHideAnimation={disableHideAnimation}
-			disableContentHover={disableContentHover}
-			disableCloseOnOutsideClick={disableCloseOnOutsideClick}
-			contentGap={contentGap}
-			containerRef={containerRef}
 			contentClassName={contentClassName}
 			contentAttributes={{ style: { ...paddingStyles?.variables } }}
-		>
-			{children}
-		</Flyout>
+		/>
 	);
 };