+++ src/styles/types.ts
@@ -2,6 +2,8 @@ import React from "react";
 import * as G from "types/global";
 
 export type TextAlign = "start" | "center" | "end";
+export type Justify = "start" | "center" | "end" | "space-between";
+export type Align = "start" | "center" | "end" | "stretch" | "baseline";
 export type Radius = "small" | "medium" | "large" | "circular" | "none";
 export type Position = "relative" | "absolute" | "fixed" | "sticky" | "static";
 export type BorderColor =