+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 3.2.0
 
+- Grid utility
+
 - Theming: Updated rem calculations to be based on 16px
 - Theming: Updated generated bg neutral faded values (slightly more faded)
 
@@ -40,3 +42,8 @@
 - Loader: Large size support
 - Flyout: fixed updatePosition not working when flyout stays opened
 - Flyout: better fallback calculations and correctly calculating the gap when positioning
+- Flyout, Popover, DropdownMenu: deprecated forcePosition, added fallbackPositions array and false
+- Flyout: fallback history when no position fits
+- Flyout, Popover: initialFocusRef
+- View: removed additional div around the content when it has a single child
+- Flyout: Optimized resize observer to only be active when rendered